ci: remove ubuntu-20.04 runner builds on GHA
The Ubuntu 20.04 image on GitHub actions will be unavailable by
2025-04-15. See <https://github.com/actions/runner-images/issues/11101>
for more information on the deprecation and removal.

Therefore, the GCC 8 build on Ubuntu 20.04 is removed, but builds
with GCC 12, 13 and 14 on Ubuntu 24.04 are added. Furthermore,
the Clang 7, 8, 9 and 10 builds on Ubuntu 20.04 are removed.
Builds with newer Clang versions are added instead.